backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.129.2
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-jose==3.3.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
